Vulkan and OpenGL fixes & added basic material system

+++ Eklipse/src/Eklipse/Platform/Vulkan/VKDescriptor.cpp
@@ -83,7 +83,7 @@ namespace Eklipse
 		{
 			VkDescriptorSetLayoutCreateInfo layoutInfo{};
 			layoutInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-			layoutInfo.bindingCount = bindings.size();
+			layoutInfo.bindingCount = static_cast<uint32_t>(bindings.size());
 			layoutInfo.pBindings = bindings.data();
 
 			VkDescriptorSetLayout layout;
